fixed bugs and added a toggle

+++ AntiRacism.py
@@ -19,7 +19,36 @@ from telethon.tl.types import Message
 
 
 class AntiRacismMod(loader.Module):
-    strings = {"name": "AntiRacism"}
+    """delete racists messages"""
+
+    strings = {
+        "name": "AntiRacism",
+        "enabled": "🐈 AntiRacism enabled",
+        "disabled": "🐈‍⬛ AntiRacism disabled",
+    }
+    strings_ru = {
+        "enabled": "🐈 Антирасизм включен",
+        "disabled": "🐈‍⬛ Антирасизм выключен",
+    }
+
+    async def client_ready(self, client, db):
+        self.db = db
+        self.client = client
+
+    async def antiracismcmd(self, message: Message):
+        "acivate/deactivate antiracism"
+        status = self.db.get(
+            "anti",
+            "status",
+        )
+        if status == "":
+            self.db.set("anti", "status", False)
+        if status == False:
+            self.db.set("anti", "status", True)
+            await utils.answer(message, f"<b>{self.strings('enabled')}</b>")
+        else:
+            self.db.set("anti", "status", False)
+            await utils.answer(message, f"<b>{self.strings('disabled')}</b>")
 
     async def watcher(self, message: Message):
         text = message.text
@@ -36,7 +65,20 @@ class AntiRacismMod(loader.Module):
             "нигги",
             "негрики",
             "негриков",
+            "nigga",
+            "nigger",
+            "niga",
+            "niger",
+            "niggers",
+            "nigers",
         ]
-        for i in word:
-            if i in text:
-                await message.delete()
+        status = self.db.get(
+            "anti",
+            "status",
+        )
+        if status == False:
+            return
+        if status == True:
+            for i in word:
+                if i in text:
+                    await message.delete()
